+++ __tests__/terminal.spec.js
@@ -408,6 +408,16 @@ function output(term) {
         return $(this).text().replace(/\xA0/g, ' ');
     }).get();
 }
+function no_formatters() {
+    let formatters;
+    beforeEach(() => {
+        formatters = $.terminal.defaults.formatters;
+        $.terminal.defaults.formatters = [];
+    });
+    afterEach(() => {
+        $.terminal.defaults.formatters = formatters;
+    });
+}
 function timer(callback, timeout) {
     return new Promise(function(resolve, reject) {
         setTimeout(function() {
@@ -3313,6 +3323,7 @@ describe('sub plugins', function() {
 describe('Terminal plugin', function() {
     describe('jQuery Terminal options', function() {
         describe('prompt', function() {
+            no_formatters();
             it('should have default prompt', function() {
                 var term = $('<div/>').terminal($.noop);
                 expect(term.get_prompt()).toEqual('> ');
@@ -3352,7 +3363,7 @@ describe('Terminal plugin', function() {
             it('should render raw prompt', async () => {
                 const prompt = '<span>$</span><span> </span>';
                 const prompts = [
-                    '<span>$</span> ',
+                    prompt,
                     function(fn) {
                         fn(prompt);
                     },
@@ -3374,10 +3385,72 @@ describe('Terminal plugin', function() {
                     expect(term.find('.cmd-prompt').text()).toEqual('$ ');
                 }
             });
+            it('should render raw prompt from echo command', async () => {
+                const prompt = '<span>$</span><span> </span>';
+                const prompts = [
+                    prompt,
+                    function(fn) {
+                        fn(prompt);
+                    },
+                    function() {
+                        return prompt;
+                    },
+                    async function() {
+                        return prompt;
+                    }
+                ];
+                for (const prompt of prompts) {
+                    const term = $('<div/>').terminal($.noop, {
+                        greetings: null,
+                        raw: {
+                            prompt: true
+                        },
+                        prompt
+                    });
+                    await delay(10);
+                    await term.exec('lorem');
+                    await term.exec('ipsum');
+                    expect(term.get_output().split('\n')).toEqual([
+                        '<span>$</span><span> </span>lorem',
+                        '<span>$</span><span> </span>ipsum'
+                    ]);
+                }
+            });
+            it('should render raw prompt from exec animation', async () => {
+                const prompt = '<span>$</span><span> </span>';
+                const prompts = [
+                    prompt,
+                    function(fn) {
+                        fn(prompt);
+                    },
+                    function() {
+                        return prompt;
+                    },
+                    async function() {
+                        return prompt;
+                    }
+                ];
+                for (const prompt of prompts) {
+                    const term = $('<div/>').terminal($.noop, {
+                        greetings: null,
+                        raw: {
+                            prompt: true
+                        },
+                        prompt
+                    });
+                    await delay(10);
+                    await term.exec('lorem', { typing: true, delay: 0 });
+                    await term.exec('ipsum', { typing: true, delay: 0 });
+                    expect([prompt, output(term)]).toEqual([
+                        prompt,
+                        ['$ lorem', '$ ipsum']
+                    ]);
+                }
+            });
             it('should render raw prompt with typing animation', async () => {
                 const prompt = '<span>$</span><span> </span>';
                 const prompts = [
-                    '<span>$</span> ',
+                    prompt,
                     function(fn) {
                         fn(prompt);
                     },
@@ -3582,14 +3655,9 @@ describe('Terminal plugin', function() {
             var term = $('<div/>').terminal($.noop, {
                 raw: true
             });
-            let formatters;
+            no_formatters();
             beforeEach(() => {
-                formatters = $.terminal.defaults.formatters;
-                $.terminal.defaults.formatters = [];
                 term.clear();
-            });
-            afterEach(() => {
-                $.terminal.defaults.formatters = formatters;
             });
             const img = '<img src="http://lorempixel.com/300/200/cats/"/>';
             it('should display html when no raw echo option is specified', function() {